Revert "Merge branch 'Beep6581/dev' into dev"
This reverts commit 276e61641d6b0f0de0d4d56a0b4867a656c6e7c7.

+++ rtengine/myfile.cc
@@ -364,7 +364,7 @@ int fscanf (IMFILE* f, const char* s ...)
     char buf[50], *endptr = nullptr;
     int copy_sz = f->size - f->pos;
 
-    if (copy_sz > static_cast<int>(sizeof(buf))) {
+    if (copy_sz > sizeof(buf)) {
         copy_sz = sizeof(buf) - 1;
     }
